src/test/java/strings/longest/palindromic/substring_5/SolutionTest.java
package strings.longest.palindromic.substring_5;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;


class SolutionTest {

    private Solution solution = new Solution();

    @Test
    void longestPalindromeAca() {
        String input = "aacabdkacaa";

        String result = solution.longestPalindrome(input);

        assertEquals("aca", result);
    }

    @Test
    void longestPalindromeBacab() {
        String input = "bacab";

        String result = solution.longestPalindrome(input);

        assertEquals("bacab", result);
    }

    @Test
    void longestPalindromeBb() {
        String input = "bb";

        String result = solution.longestPalindrome(input);

        assertEquals("bb", result);
    }

    @Test
    void longestPalindromeBbFromCbbd() {
        String input = "cbbd";

        String result = solution.longestPalindrome(input);

        assertEquals("bb", result);
    }

    @Test
    void longestPalindromeCcc() {
        String input = "ccc";

        String result = solution.longestPalindrome(input);

        assertEquals("ccc", result);
    }

    @Test
    void longestPalindromeBabad() {
        String input = "babad";

        String result = solution.longestPalindrome(input);

        assertTrue(result.equals("bab") || result.equals("aba"));
    }

    @Test
    void longestPalindromeAbba() {
        String input = "abba";

        String result = solution.longestPalindrome(input);

        assertEquals("abba", result);
    }

    @Test
    void longestPalindromeAaabaaaa() {
        String input = "aaabaaaa";

        String result = solution.longestPalindrome(input);

        assertEquals("aaabaaa", result);
    }

    @Test
    void longestPalindromeLongInput() {
        String input = "azwdzwmwcqzgcobeeiphemqbjtxzwkhiqpbrprocbppbxrnsxnwgikiaqutwpftbiinlnpyqstkiqzbggcsdzzjbrkf" +
                "mhgtnbujzszxsycmvipjtktpebaafycngqasbbhxaeawwmkjcziybxowkaibqnndcjbsoehtamhspnidjylyisiaewmypfyiq" +
                "twlmejkpzlieolfdjnxntonnzfgcqlcfpoxcwqctalwrgwhvqvtrpwemxhirpgizjffqgntsmvzldpjfijdncexbwtxnmbnoy" +
                "kxshkqbounzrewkpqjxocvaufnhunsmsazgibxedtopnccriwcfzeomsrrangufkjfzipkmwfbmkarnyyrgdsooosgqlkzvor" +
                "rrsaveuoxjeajvbdpgxlcrtqomliphnlehgrzgwujogxteyulphhuhwyoyvcxqatfkboahfqhjgujcaapoyqtsdqfwnijlkkn" +
                "uralezqmcryvkankszmzpgqutojoyzsnyfwsyeqqzrlhzbc";

        String result = solution.longestPalindrome(input);

        assertEquals("sooos", result);
    }

    @Test
    void longestPalindromeLongInputLongReturn() {
        String input = "aaaabbbbbbbbbbccccccccccddddddddddeeeeeeeeeeffffffffffgggggggggghhhhhhhhhhiiiiiiiiii" +
                "jjjjjjjjjjkkkkkkkkkkllllllllllmmmmmmmmmmnnnnnnnnnnooooooooooppppppppppqqqqqqqqqqrrrrrrrrrrssssssssss" +
                "ttttttttttuuuuuuuuuuvvvvvvvvvvwwwwwwwwwwxxxxxxxxxxyyyyyyyyyyzzzzzzzzzzyyyyyyyyyyxxxxxxxxxxwwwwwwwwww" +
                "vvvvvvvvvvuuuuuuuuuuttttttttttssssssssssrrrrrrrrrrqqqqqqqqqqppppppppppoooooooooonnnnnnnnnnmmmmmmmmmm" +
                "llllllllllkkkkkkkkkkjjjjjjjjjjiiiiiiiiiihhhhhhhhhhggggggggggffffffffffeeeeeeeeeeddddddddddcccccccccc" +
                "bbbbbbbbbbaaaaaaaabbbbbbbbbbccccccccccddddddddddeeeeeeeeeeffffffffffgggggggggghhhhhhhhhhiiiiiiiiii" +
                "jjjjjjjjjjkkkkkkkkkkllllllllllmmmmmmmmmmnnnnnnnnnnooooooooooppppppppppqqqqqqqqqqrrrrrrrrrrssssssssss" +
                "ttttttttttuuuuuuuuuuvvvvvvvvvvwwwwwwwwwwxxxxxxxxxxyyyyyyyyyyzzzzzzzzzzyyyyyyyyyyxxxxxxxxxxwwwwwwwwww" +
                "vvvvvvvvvvuuuuuuuuuuttttttttttssssssssssrrrrrrrrrrqqqqqqqqqqppppppppppoooooooooonnnnnnnnnnmmmmmmmmmm" +
                "llllllllllkkkkkkkkkkjjjjjjjjjjiiiiiiiiiihhhhhhhhhhggggggggggffffffffffeeeeeeeeeeddddddddddcccccccccc" +
                "bbbbbbbbbbaaaa";

        String result = solution.longestPalindrome(input);

        assertEquals("aaaabbbbbbbbbbccccccccccddddddddddeeeeeeeeeeffffffffffgggggggggghhhhhhhhhhiiiiiiiiiijjjjjjjjjjkkkkkkkkkkllllllllllmmmmmmmmmmnnnnnnnnnnooooooooooppppppppppqqqqqqqqqqrrrrrrrrrrssssssssssttttttttttuuuuuuuuuuvvvvvvvvvvwwwwwwwwwwxxxxxxxxxxyyyyyyyyyyzzzzzzzzzzyyyyyyyyyyxxxxxxxxxxwwwwwwwwwwvvvvvvvvvvuuuuuuuuuuttttttttttssssssssssrrrrrrrrrrqqqqqqqqqqppppppppppoooooooooonnnnnnnnnnmmmmmmmmmmllllllllllkkkkkkkkkkjjjjjjjjjjiiiiiiiiiihhhhhhhhhhggggggggggffffffffffeeeeeeeeeeddddddddddccccccccccbbbbbbbbbbaaaaaaaabbbbbbbbbbccccccccccddddddddddeeeeeeeeeeffffffffffgggggggggghhhhhhhhhhiiiiiiiiiijjjjjjjjjjkkkkkkkkkkllllllllllmmmmmmmmmmnnnnnnnnnnooooooooooppppppppppqqqqqqqqqqrrrrrrrrrrssssssssssttttttttttuuuuuuuuuuvvvvvvvvvvwwwwwwwwwwxxxxxxxxxxyyyyyyyyyyzzzzzzzzzzyyyyyyyyyyxxxxxxxxxxwwwwwwwwwwvvvvvvvvvvuuuuuuuuuuttttttttttssssssssssrrrrrrrrrrqqqqqqqqqqppppppppppoooooooooonnnnnnnnnnmmmmmmmmmmllllllllllkkkkkkkkkkjjjjjjjjjjiiiiiiiiiihhhhhhhhhhggggggggggffffffffffeeeeeeeeeeddddddddddccccccccccbbbbbbbbbbaaaa", result);
    }

}